fix: fix saveChat return value

+++ Back/paperdot/src/main/java/gomgook/paperdot/chat/dto/ChatMessageDto.java
@@ -10,7 +10,7 @@ import lombok.*;
 @NoArgsConstructor
 public class ChatMessageDto {
 
-    private Long chatRoomId;
+    private int chatRoomId;
     private Long senderId;
     private String nickname;
     private String message;
